requirements.txt
h5py==3.1.0
matplotlib==3.1.2
numpy==1.17.4
pandas==1.3.4
Pillow==9.1.1
PyQt5==5.15.6
PyYAML==6.0
